rename | reDirect.js
script 내에 있던 reDirect 함수를 모듈로 분리합니다.
이 모듈은 다른 페이지로 움직이는 행동과 로딩 바의 애니메이션을 담당합니다
